+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 4.5.0 (2021-06-07)
+
+- Added experimental support for OpenSearch
+- Added support for synonyms in Japanese
+
+## 4.4.4 (2021-03-12)
+
+- Fixed `too_long_frame_exception` with `scroll` method
+- Fixed multi-word emoji tokenization
+
+## 4.4.3 (2021-02-25)
+
+- Added support for Hunspell
+- Fixed warning about accessing system indices
+
 ## 4.4.2 (2020-11-23)
 
 - Added `missing_records` method to results